Use db.User for api internal operations
Instead of using the api.User, use directly the database user.

The api user only uses one location that can be obfuscated or not. This caused errors when searching by distance, because, it was using the obfuscated location instead of the normal location when performing the search, causing the center of the search to move and showing different results on each test run.

On this way, we ensure to send only user with obfuscated location when the user data is on the response, and use the DB user inside the API functions instead, to ensure have all the correct data.